CLI-623 Address review: relocate finalizeAgentInstall, share CAG skill via .agents/skills
- Move finalizeAgentInstall out of agent-integrate-prelude (opening steps) into
  a dedicated agent-integrate-install.ts (closing step); update cursor/codex/
  copilot imports. The prelude file is back to preflight-only concerns.
- Write the Cursor CAG skill to the shared .agents/skills/sonar-context-augmentation/
  SKILL.md instead of a Cursor-private .cursor/skills copy. Cursor reads
  .agents/skills (same as Codex and Antigravity), so the three tools share one
  skill rather than duplicating the same skill name across two directories.
