+++ src/cli/commands/integrate/codex/index.ts
@@ -21,10 +21,8 @@
 // Integrate command — setup SonarQube integration for Codex.
 
 import type { ResolvedAuth } from '../../../../lib/auth-resolver';
-import {
-  displayAgentIntegratePrelude,
-  finalizeAgentInstall,
-} from '../_common/agent-integrate-prelude';
+import { finalizeAgentInstall } from '../_common/agent-integrate-install';
+import { displayAgentIntegratePrelude } from '../_common/agent-integrate-prelude';
 import { resolveSqaaSetup } from '../_common/sqaa-entitlement';
 import type { IntegrateAgentOptions } from '../_common/types';
 import { CODEX_INTEGRATION_ID, type CodexIntegrationOptions } from './declaration';
