+++ src/cli/commands/integrate/copilot/index.ts
@@ -19,10 +19,8 @@
  */
 
 import type { ResolvedAuth } from '../../../../lib/auth-resolver';
-import {
-  displayAgentIntegratePrelude,
-  finalizeAgentInstall,
-} from '../_common/agent-integrate-prelude';
+import { finalizeAgentInstall } from '../_common/agent-integrate-install';
+import { displayAgentIntegratePrelude } from '../_common/agent-integrate-prelude';
 import { resolveSqaaSetup } from '../_common/sqaa-entitlement';
 import type { IntegrateAgentOptions } from '../_common/types';
 import { COPILOT_INTEGRATION_ID, type CopilotIntegrationOptions } from './declaration';
